dcp-815 separate class for metadata/descriptor.py
